.github/scripts/requirements.txt
pyyaml>=6.0
GitPython>=3.1.30
PyGithub>=1.57
